requirements.txt
couchdb
python-dotenv
